问题模块 代码自测&测评实现

+++ yicode-question/yicode-question-api/src/main/java/com/yixihan/yicode/question/api/dto/request/question/AddQuestionAnswerDtoReq.java
@@ -37,5 +37,5 @@ public class AddQuestionAnswerDtoReq {
     private Integer answerTimeConsume;
     
     @ApiModelProperty(value = "内存消耗")
-    private Integer answerMemoryConsume;
+    private Double answerMemoryConsume;
 }
